Tests: rename utils files
